refactor(bff): convert ExpiredSessionCleanup to IBackgroundJob (#627)
Replace static handler with proper IBackgroundJob record +
RecurringJob attribute. BffExpiredSessionCleanupJob runs every 15 min.
Added Granit.BackgroundJobs dependency to Bff.EntityFrameworkCore.
Handler follows Wolverine convention (job record as first param).
