+++ .github/shard-filters/architecture.slnf
@@ -24,6 +24,7 @@
       "src/Granit.Authentication.GoogleCloud/Granit.Authentication.GoogleCloud.csproj",
       "src/Granit.Authentication.JwtBearer/Granit.Authentication.JwtBearer.csproj",
       "src/Granit.Authentication.Keycloak/Granit.Authentication.Keycloak.csproj",
+      "src/Granit.Authentication.Oidc/Granit.Authentication.Oidc.csproj",
       "src/Granit.Authentication.OpenIddict/Granit.Authentication.OpenIddict.csproj",
       "src/Granit.Authorization.AI/Granit.Authorization.AI.csproj",
       "src/Granit.Authorization.Endpoints/Granit.Authorization.Endpoints.csproj",
