+++ src/Granit.Bff.EntityFrameworkCore/Granit.Bff.EntityFrameworkCore.csproj
@@ -16,6 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Granit.BackgroundJobs\Granit.BackgroundJobs.csproj" />
     <ProjectReference Include="..\Granit.Bff\Granit.Bff.csproj" />
     <ProjectReference Include="..\Granit.Guids\Granit.Guids.csproj" />
     <ProjectReference Include="..\Granit.Persistence\Granit.Persistence.csproj" />
